requirements.txt
fastapi~=0.115.8
pydantic~=2.10.6
mysql-connector-python~=9.2.0
streamlit~=1.42.0
pandas~=2.2.0
requests~=2.32.3